Remove AsyncLocal<T> light-up when the target platform guarantees it is present
