+++ src/Microsoft.VisualStudio.Threading/LightUps`1.cs
@@ -19,18 +19,20 @@ namespace Microsoft.VisualStudio.Threading
     /// <typeparam name="T">The generic type argument.</typeparam>
     internal static class LightUps<T>
     {
-#if !NET46
+#if TRYSETCANCELEDCT
+        /// <summary>
+        /// A delegate that invokes <see cref="TaskCompletionSource{TResult}.TrySetCanceled(CancellationToken)"/>.
+        /// </summary>
+#else
         /// <summary>
         /// A delegate that invokes the <see cref="TaskCompletionSource{TResult}.TrySetCanceled"/>
         /// method that takes <see cref="CancellationToken"/> as an argument.
         /// Will be <c>null</c> on .NET Framework versions under 4.6.
         /// </summary>
-#else
-        /// <summary>
-        /// A delegate that invokes <see cref="TaskCompletionSource{TResult}.TrySetCanceled(CancellationToken)"/>.
-        /// </summary>
 #endif
         internal static readonly Func<TaskCompletionSource<T>, CancellationToken, bool> TrySetCanceled;
+
+#if !ASYNCLOCAL
 
         /// <summary>
         /// A value indicating whether the BCL AsyncLocal{T} type is available.
@@ -51,6 +53,8 @@ namespace Microsoft.VisualStudio.Threading
         /// </summary>
         private static readonly PropertyInfo BclAsyncLocalValueProperty;
 
+#endif
+
         /// <summary>
         /// Initializes static members of the <see cref="LightUps{T}"/> class.
         /// </summary>
@@ -67,14 +71,18 @@ namespace Microsoft.VisualStudio.Threading
                     TrySetCanceled = (Func<TaskCompletionSource<T>, CancellationToken, bool>)methodInfo.CreateDelegate(typeof(Func<TaskCompletionSource<T>, CancellationToken, bool>));
                 }
 
+#if !ASYNCLOCAL
                 if (LightUps.BclAsyncLocalType != null)
                 {
                     BclAsyncLocalType = LightUps.BclAsyncLocalType.MakeGenericType(typeof(T));
                     BclAsyncLocalValueProperty = BclAsyncLocalType.GetTypeInfo().GetDeclaredProperty("Value");
                     IsAsyncLocalSupported = true;
                 }
+#endif
             }
         }
+
+#if !ASYNCLOCAL
 
         /// <summary>
         /// Creates an instance of the BCL AsyncLocal{T} type.
@@ -168,5 +176,6 @@ namespace Microsoft.VisualStudio.Threading
             /// <inheritdoc />
             internal override object CreateAsyncLocal() => new TAsyncLocal();
         }
+#endif
     }
 }
